Milestone_3/Webshop/Backend_django/requirements.txt
asgiref==3.8.1
Django==4.2
psycopg2-binary==2.9.10
sqlparse==0.5.3
tzdata==2025.1
dj-database-url
djangorestframework
django-cors-headers
python-dotenv
Pillow
django-filter
stripe